move TestHouseholds to correct path
